requirements.txt
tensorflow-gpu==2.4.0
keras==2.3.1
scipy
numpy==1.19.2
pandas
scikit-learn>=0.18.2
h5py
matplotlib
joblib>=0.12
scikit-image
